Move `PrimitiveValType::lower_gc` into the `component_types` submodule
